finished the navbar with all fixed
