+++ src/Components/Layout.tsx
@@ -1,4 +1,4 @@
-import { Navbar } from './Navbar';
+import { Navbar } from './Navbar/Navbar';
 import { Outlet } from 'react-router-dom';
 
 interface LayoutProps {
